chore: wip

+++ src/controllers/category.controller.ts
@@ -1,10 +1,13 @@
 //category
 import { Request, Response, NextFunction } from 'express';
 import CategoryService from '../services/category.service';
+import ApiError from '../error/apiError';
 
 class CategoryController {
   async createCategory(request: Request, response: Response, next: NextFunction) {
     try {
+      if (!request.body.name) throw ApiError.BadRequest("name");
+
       const name = request.body.name;
       const result = await CategoryService.createCategory(name);
 
@@ -19,7 +22,9 @@ class CategoryController {
 
   async findCategoryById(request: Request, response: Response, next: NextFunction) {
     try {
-      const id = request.params.id as unknown as number;
+      if (!request.body.id) throw ApiError.BadRequest("id");
+
+      const id = +request.params.id;
       const result = await CategoryService.findCategoryById(id);
       console.log(result);
 
@@ -34,7 +39,9 @@ class CategoryController {
 
   async deleteCategoryById(request: Request, response: Response, next: NextFunction) {
     try {
-      const id = request.params.id as unknown as number;
+      if (!request.body.id) throw ApiError.BadRequest("id");
+
+      const id = +request.params.id;
       await CategoryService.deleteCategoryById(id);
       return response.status(200).send({
         success: true,
@@ -50,7 +57,7 @@ class CategoryController {
   //todo
   async findGoodsByCategoryId(request: Request, response: Response, next: NextFunction) {
     try {
-      const id = request.params.id as unknown as number;
+      const id = +request.params.id;
       const result = await CategoryService.findGoodsByCategoryId(id);
 
       return response.status(200).send(result);
@@ -61,8 +68,8 @@ class CategoryController {
 
   async updateCategoryById(request: Request, response: Response, next: NextFunction) {
     try {
-      const name = request.body.name;
-      const id = request.params.id as unknown as number;
+      const name: string = request.body.name;
+      const id = +request.params.id;
       const input = { id: id, name: name };
       const result = await CategoryService.updateCategoryById(id, input);
